resi piu' robusti rispetto al paper loglog e hll e aggiunto notebook per la differenza tra hll e hllpp

+++ src/satp/algorithms/HyperLogLog.h
@@ -12,7 +12,9 @@ using namespace std;
 namespace satp::algorithms {
     class HyperLogLog final : public Algorithm {
     public:
-        // k = number of register index bits (m = 2^k registers)
+        // Paper-strict mode (Flajolet et al. 2007 practical setting):
+        // - k in [4,16], m = 2^k registers
+        // - 32-bit hash domain (L = 32)
         explicit HyperLogLog(uint32_t K, uint32_t L);
 
         void process(uint32_t id) override;
@@ -26,8 +28,9 @@ namespace satp::algorithms {
     private:
         uint32_t k;
         uint32_t numberOfBuckets; // nel paper coincide con m
-        uint32_t lengthOfBitMap; // lunghezza della bitmap, valore inutile, perché il paper assume che
-        vector<uint32_t> bitmap; // bitmap
+        uint32_t lengthOfBitMap;
+        // Logical register values fit in <=5 bits in paper settings; we store in uint8_t.
+        vector<uint8_t> bitmap;
         double alphaM;
         double sumInversePowers; // \sum_j 2^{-M[j]}
         uint32_t zeroRegisters;
