+++ src/satp/algorithms/LogLog.h
@@ -12,7 +12,9 @@ using namespace std;
 namespace satp::algorithms {
     class LogLog final : public Algorithm {
     public:
-        // k = number of register index bits (m = 2^k registers)
+        // Paper-strict mode (Durand-Flajolet 2003 / HLL 2007 practical range):
+        // - k in [4,16], m = 2^k registers
+        // - 32-bit hash domain (L = 32)
         explicit LogLog(uint32_t K, uint32_t L);
 
         void process(uint32_t id) override;
@@ -27,7 +29,8 @@ namespace satp::algorithms {
         uint32_t k;
         uint32_t numberOfBuckets;
         uint32_t lengthOfBitMap;
-        vector<uint32_t> bitmap;
+        // Logical register values fit in <=5 bits in paper settings; we store in uint8_t.
+        vector<uint8_t> bitmap;
         double sumRegisters; // \sum_j M[j]
 
         static constexpr double ALPHA_INF = 0.39701;
